Adding speeddating event config and fixing integration w/Eventconfig*

+++ component/admin/sql/install.mysql.sql
@@ -224,7 +224,7 @@ CREATE TABLE IF NOT EXISTS `#__claw_packages` (
   `published` TINYINT(4) DEFAULT NULL,
   `description` VARCHAR(255) NOT NULL DEFAULT '',
   `alias` VARCHAR(255) NOT NULL DEFAULT '',
-  `eventPackageType` TINYINT(4) NOT NULL DEFAULT 0,
+  `eventPackageType` SMALLINT(5) NOT NULL DEFAULT 0,
   `packageInfoType` TINYINT(4) NOT NULL DEFAULT 0,
   `couponKey` VARCHAR(10) NOT NULL DEFAULT '',
   `couponValue` FLOAT NOT NULL DEFAULT 0.0,
